Extract platform independent code
